move code to "net.h" and "eth-stm32f7.h"
